refactor: re-usable
